# Remove F Monad from Markdown and Org readers.