--- a/src/Text/Pandoc/Readers/Markdown.hs
+++ b/src/Text/Pandoc/Readers/Markdown.hs
@@ -1292,7 +1292,7 @@ pipeBreak = try $ do
 pipeTable :: MarkdownParser ([Alignment], [Double], [Blocks], [[Blocks]])
 pipeTable = try $ do
   (heads,aligns) <- (,) <$> pipeTableRow <*> pipeBreak
-  lines' <- sequence <$> many1 pipeTableRow
+  lines' <-  many1 pipeTableRow
   let widths = replicate (length aligns) 0.0
   return (aligns, widths, heads, lines')