@tarleb 

Note that these two modules now contain circular definitions. Normal editing (not touching the footnote handling) should be fine but there are some points where care has to be exercised to avoid loops.

One particular point is the different semantics of

```
if foo 
  then
    return True
  else
    return False
```

and

```
return $ if foo then True else False
```

The second is lazier than the first as `foo` must be evaluated in the first instance but not in the second.