@jgm 2015-02-17
+++ Matthew Pickering [Feb 16 15 15:25 ]:
>   Note that these two modules now contain circular definitions. Normal

Can you elaborate?