@mpickering 2015-02-18
Here is an example which demonstrates how the reader now resolve footnotes. 

```
module Example where

import Text.Parsec
import Control.Monad.Reader

test1 :: ParsecT String [(Char, String)] (Reader [(Char, String)]) String
test1 = do
    a <- anyChar
    notes <- ask
    -- It is important that notes is not forced
    return $ case lookup a notes of
        Just v -> v
        Nothing -> ""

test1a :: (Char, String) -> ParsecT String [(Char, String)] (Reader [(Char, String)]) ()
test1a v = 
    updateState (v:)

prog :: ParsecT String [(Char, String)] (Reader [(Char, String)]) String
prog = do
    res <- test1
    res2 <- test1
    test1a ('a',"1")
    test1a ('b',"2")
    return (res ++ res2)

main :: IO ()
main = print runTest2

retState :: Monad m => ParsecT s st m a -> ParsecT s st m (a, st)
retState p = do
    res <- p
    st <- getState
    return (res, st)


runTest2 :: (String, [(Char, String)])
runTest2 = (res, n)
    where
        m  = runParserT (retState prog) [] "" "abc"
        -- Notice how this is circular
        (res, n) = either undefined id $ runReader m n
```