@mpickering 2015-02-18
By default Haskell's evaluation strategy is lazy, that is, values are not computed before they are required. Notice that when we parse a footnote marker that we will definitely replace it with something (either a `Note` or `Str "[1^]"`, the only problem is that we don't know which until we reach the bottom of the file. This solution makes sure evaluation is delayed long enough so that by the time we force evaluation, we have all the necessary information to fill in the gaps.

The final line looks a bit mystical but there is no cycle as the state is not modified as a result of any reader actions. In fact, a very common pattern is the following.

```
st <- ask
return $ ...
```

Which ensures that the value of `st` is not forced. I'm not sure if this is a very good explanation, please ask more questions if anything is unclear.

Benchmarks

old
```
benchmarking markdown reader
mean: 55.39495 ms, lb 53.88569 ms, ub 57.09139 ms, ci 0.950
std dev: 3.738574 ms, lb 3.045484 ms, ub 4.593733 ms, ci 0.950
variance introduced by outliers: 28.456%
variance is moderately inflated by outliers
```

new
```
benchmarking markdown reader
mean: 70.94543 ms, lb 69.42924 ms, ub 73.30225 ms, ci 0.950
std dev: 4.368495 ms, lb 2.901871 ms, ub 6.838932 ms, ci 0.950
found 1 outliers among 20 samples (5.0%)
  1 (5.0%) high mild
variance introduced by outliers: 23.712%
variance is moderately inflated by outliers
```