@jgm 2015-03-15
So there seem to be two disadvantages of this change:

- We have to be careful about laziness or things can blow up
- Performance is worse

What would you say are the advantages?  How should we weigh them against the disadvantages?