@mpickering 2015-03-16
Here are a few reasons.

* Reduces cognitive overhead. I always found it difficult to work out how many times to `return` when making changes to the markdown reader. You often reach for functions such as `trimInlines` but then realise you need to use `trimInlinesF` which just disrupts the normal writing process. 
* Improves code reusability (for example [this](https://github.com/jgm/pandoc/blob/master/src/Text/Pandoc/Readers/Markdown.hs#L1931) can be reused from `Parsing`)
* Likely to reduce bugs (due to things like [this](https://github.com/jgm/pandoc/blob/master/src/Text/Pandoc/Readers/Markdown.hs#L494) and general clarity)
* Reduces code duplication with functions defined for `F` and normally.

As for the disadvantages.

1. You only have to be careful about laziness if you are making changes to the part of the program which relies on laziness. (Which is a small part!) This is in contrast to currently where you always have to think about `F` even in parts which don't concern it.
2. I'm not too concerned about performance implications. I can look at profiling the reader to find other places to improve if you feel the hit is unacceptable.