@jgm 2015-03-16
@mpickering, this PR can't be automatically merged due to subsequent changes in the Org and Markdown readers.  Since you understand this code better than I, could you rebase your changes on top of current HEAD?