@mpickering 2015-03-22
Rebased, just waiting for travis now.